Add an abstract mother class for breadth-first searches.
